refactor:muda nome do controller a StudentWorkoutController
Apaga Try e Catch
