+++ routes/api.php
@@ -3,7 +3,7 @@
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\StudentController;
 use App\Http\Controllers\UserController;
-use App\Http\Controllers\WorkoutController;
+use App\Http\Controllers\StudentWorkoutController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -28,7 +28,7 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::post('students', [StudentController::class, 'store'])->middleware(['ability:create-students']);
     //Route::get('workouts', [WorkoutController::class, 'index'])->middleware(['ability:get-workouts']);
     Route::post('logout', [AuthController::class, 'logout']);
-    Route::get('students/{id}/workouts', [WorkoutController::class, 'workoutsByStudent'])->middleware(['ability:get-workout']);
+    Route::get('students/{id}/workouts', [StudentWorkoutController::class, 'workoutsByStudent'])->middleware(['ability:get-workout']);
 });
 
 Route::post('login', [AuthController::class, 'store']);
